First commit with code, etc.
Currently supports the following interpolation methods:
Lagrange/Newton form polynomial interpolation (these return essentially the same result, but get there in different ways)
Hermite-Fejer polynomial interpolation assuming 0 slope at each point
Natural cubic spline interpolation

For each of these methods, it provides the coefficients and can evaluate the interpolant on a relevant domain.

Currently, it only interpolates when given a function. However, I intend on extending this to points so that the user doesn't have to define an arbitrary function for their set of data points.

Another goal is to create graphs and possibly animations that display how sensitive each interpolant is to perturbations in the provided data.

Also, I'd like to split this up into sub-packages, because it's already getting to be a mess.